feat: Account creation works

+++ scripts/local/start-in-docker.sh
@@ -1,6 +1,10 @@
+#!/bin/sh
+#Usage exmaple
+#sudo ./scripts/local/start-in-docker.sh juno16g2rahf5846rxzp3fwlswy08fz8ccuwk03k57y  -no
+set -e
+
 # script for lauching local juno network
-if [ "$1" = "" ]
-then
+if [ "$1" = "" ]; then
   echo "Usage: $0 1 arg required - rules address"
   exit
 fi
@@ -15,26 +19,64 @@ DIR_NAME=$(basename "$PWD")
 IMAGE_NAME="juno-node-1"
 DIR_NAME_SNAKE=$(echo $DIR_NAME | tr '-' '_')
 WASM="artifacts/$DIR_NAME_SNAKE.wasm"
+STAKE_TOKEN=ujunox
+STAKE=${STAKE_TOKEN:-ustake}
+# Reset
+NoColor='\033[0m'       # Text Reset
+# Regular Colors
+Black='\033[0;30m'        # Black
+Red='\033[0;31m'          # Red
+Green='\033[0;32m'        # Green
+Yellow='\033[0;33m'       # {Yellow}
+Blue='\033[0;34m'         # Blue
+Purple='\033[0;35m'       # Purple
+Cyan='\033[0;36m'         # Cyan
+White='\033[0;37m'        # White
+
 
 # build optimized binary if it doesn't exist
 if [ ! -f "$WASM" ]; then
   echo "building optimized binary..."
   docker run --rm -v "$(pwd)":/code \
-  --mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
-  --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
-  cosmwasm/rust-optimizer:0.12.6
+    --mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
+    --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
+    cosmwasm/rust-optimizer:0.12.6
 fi
 
-# stop docker container
-cd $JUNO_DIR
-echo "stopping container..."
-docker compose down
-# delete docker container
-echo "deleting container"
-docker rm -f $IMAGE_NAME 2> /dev/null
-# build new docker container
-echo "starting local network"
-STAKE_TOKEN=STAKE UNSAFE_CORS=true docker compose up -d
+#Initialize docker yes/no
+if [ "$2" = "-yes" ]; then
+  # stop docker container
+  cd $JUNO_DIR
+  echo "stopping container..."
+  docker compose down
+  # delete docker container
+  echo "deleting container"
+  docker rm -f $IMAGE_NAME 2>/dev/null
+  # build new docker container
+  echo "${Yellow}Starting local network${NoColor}"
+  STAKE_TOKEN=ujunox UNSAFE_CORS=true docker-compose up -d
+  echo "Adding new users..."
+
+  # add new users
+  ALICE_SEED=$($BINARY keys mnemonic)
+  echo $ALICE_SEED | $BINARY keys add alice --recover
+
+  BOB_SEED=$($BINARY keys mnemonic)
+  echo $BOB_SEED | $BINARY keys add bob --recover
+
+  OWNER_SEED=$($BINARY keys mnemonic)
+  echo $OWNER_SEED | $BINARY keys add owner --recover
+
+  AGENT_SEED=$($BINARY keys mnemonic)
+  echo $AGENT_SEED | $BINARY keys add agent --recover
+
+  USER_SEED=$($BINARY keys mnemonic)
+  echo $USER_SEED | $BINARY keys add user --recover
+
+
+  sleep 10
+fi
+
 # move binary to docker container
 cd $DIR
 docker cp "artifacts/$DIR_NAME_SNAKE.wasm" "$IMAGE_NAME:/$DIR_NAME_SNAKE.wasm"
@@ -42,36 +84,68 @@ cd $JUNO_DIR
 
 # wait for chain starting before contract storing
 
-# add new users
-ALICE_SEED=$(junod keys mnemonic)
-echo $ALICE_SEED | $BINARY keys add alice --recover
-ALICE_ADDR=$($BINARY keys show alice --address)
-
-BOB_SEED=$(junod keys mnemonic)
-echo $BOB_SEED | $BINARY keys add bob --recover
-BOB_ADDR=$($BINARY keys show bob --address)
-
-
 # send them some coins
 VALIDATOR_ADDR=$($BINARY keys show validator --address)
-echo "Validator -" $VALIDATOR_ADDR
+VALIDATOR_BALANCE=$($BINARY q bank balances $($BINARY keys show validator --address))
+echo "${Cyan}Validator :" $VALIDATOR_ADDR $"{NoColor}"
+echo "${Green}Validator Balance :" $VALIDATOR_BALANCE $"{NoColor}"
+
+ALICE_ADDR=$($BINARY keys show alice --address)
+BOB_ADDR=$($BINARY keys show bob --address)
+OWNER_ADDR=$($BINARY keys show owner --address)
+AGENT_ADDR=$($BINARY keys show agent --address)
+USER_ADDR=$($BINARY keys show user --address)
+echo "${Cyan}Alice :" $ALICE_ADDR "${NoColor}"
+echo "${Cyan}Bob :" $BOB_ADDR "${NoColor}"
+echo "${Cyan}Owner :" $OWNER_ADDR "${NoColor}"
+echo "${Cyan}User :" $USER_ADDR "${NoColor}"
+echo "${Cyan}Agent :" $AGENT_ADDR "${NoColor}"
+
+
 # errors from this point
-set -e
 
 #$BINARY tx staking create-validator --chain-id $CHAIN_ID --from=validator --fees=0.025juno --pubkey=$VALIDATOR_ADDR --commission-rate=0.1 --amount=100000000ujuno --moniker=vname
 #$BINARY gentx validator 15000000STAKE --chain-id $CHAIN_ID
 #$BINARY collect-gentxs
 
-$BINARY tx bank send $VALIDATOR_ADDR $ALICE_ADDR "25000000STAKE" --from "validator"  --yes --broadcast-mode block --sign-mode direct --chain-id $CHAIN_ID
-$BINARY tx bank send $VALIDATOR_ADDR $BOB_ADDR "25000000STAKE" --from "validator"  --yes --broadcast-mode block --sign-mode direct --chain-id $CHAIN_ID
+echo "${Yellow}Sending funds to users...${NoColor}"
+
+$BINARY tx bank send $VALIDATOR_ADDR $ALICE_ADDR "1$STAKE" --from validator --yes --broadcast-mode block --sign-mode direct --chain-id $CHAIN_ID
+$BINARY tx bank send $VALIDATOR_ADDR $BOB_ADDR "1$STAKE" --from validator --yes --broadcast-mode block --sign-mode direct --chain-id $CHAIN_ID
+$BINARY tx bank send $VALIDATOR_ADDR $OWNER_ADDR "60000000$STAKE" --from validator --yes --broadcast-mode block --sign-mode direct --chain-id $CHAIN_ID
+$BINARY tx bank send $VALIDATOR_ADDR $AGENT_ADDR "2000000$STAKE" --from validator --yes --broadcast-mode block --sign-mode direct --chain-id $CHAIN_ID
+$BINARY tx bank send $VALIDATOR_ADDR $USER_ADDR "2000000$STAKE" --from validator --yes --broadcast-mode block --sign-mode direct --chain-id $CHAIN_ID
+
+sleep 2
+echo "${Cyan}Funds sent...${NoColor}"
+
+ALICE_BALANCE=$($BINARY q bank balances $($BINARY keys show alice --address))
+echo "${Green}Alice Balance :" $ALICE_BALANCE "${NoColor}"
+BOB_BALANCE=$($BINARY q bank balances $($BINARY keys show bob --address))
+echo "${Green}Bob Balance :" $BOB_BALANCE "${NoColor}"
+OWNER_BALANCE=$($BINARY q bank balances $($BINARY keys show owner --address))
+echo "${Green}Owner Balance :" $OWNER_BALANCE "${NoColor}"
+AGENT_BALANCE=$($BINARY q bank balances $($BINARY keys show agent --address))
+echo "${Green}Agent Balance :" $AGENT_BALANCE "${NoColor}"
+USER_BALANCE=$($BINARY q bank balances $($BINARY keys show user --address))
+echo "${Green}User Balance :" $USER_BALANCE "${NoColor}"
+
+#---------------------------------------------------------------------------
+echo "${Yellow}Starting grpc...${NoColor}"
+# Start the Juno chain, making sure we have gRPC
+$BINARY start --grpc.address "127.0.0.1:9090" >/dev/null 2>&1 < /dev/null &
+sleep 3
+echo "${Green}Starting grpc done!${NoColor}"
+
+#---------------------------------------------------------------------------
+echo "${Yellow}Instantiating smart contract...${NoColor}"
 CONTRACT_CODE=$($BINARY tx wasm store "/$DIR_NAME_SNAKE.wasm" --from validator $TXFLAG --output json | jq -r '.logs[0].events[-1].attributes[0].value')
+echo "${Cyan}Instantiating smart contract done!${NoColor}"
+echo "${Cyan}CONTRACT_CODE :" $CONTRACT_CODE "${NoColor}"
 
-#---------- SMART CONTRACT INTERACTION ------------------------
-
-# instantiate smart contract
-INIT='{"denom":"stake","cw_rules_addr":"$cw_rules_addr"}'
-$BINARY tx wasm instantiate $CONTRACT_CODE "$INIT" --from "alice" --label "my first contract" $TXFLAG --no-admin
+INIT='{"denom":"$STAKE","cw_rules_addr":"$cw_rules_addr"}'
+$BINARY tx wasm instantiate $CONTRACT_CODE "$INIT" --from alice --label "croncat" $TXFLAG --no-admin -y
 
 # get smart contract address
 CONTRACT_ADDRESS=$($BINARY query wasm list-contract-by-code $CONTRACT_CODE --output json | jq -r '.contracts[-1]')
-
+echo "Instantiating Smart Contract Done!"
